Test fix and security for test refactor
