+++ src/main/java/com/example/petbuddybackend/dto/chat/ChatMessageDTO.java
@@ -2,8 +2,6 @@ package com.example.petbuddybackend.dto.chat;
 
 import com.example.petbuddybackend.utils.time.TimeUtils;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.validation.constraints.DecimalMax;
-import jakarta.validation.constraints.Digits;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
